Added the ability to shuffle the order beams are evaluated.
This should allow multiple processes to cooperate on rating.
Mutual exclusion is handled by the database, which will report
already-rated candidates as such. If two processes happen to rate
the same candidate at the same time, at worst you will have a
duplicate value for the rating in the database.